Code cleanup and minor changes
- ModLE now compiles with -DWARNINGS_AS_ERRORS=ON on GCC 10 and Clang 11
- suppress_compiler_warnings.hpp contains few macros to reduce the
  amount of boilerplate required to suppress compiler warnings
- Reduced the frequency to static_casts used to convert between numeric types
- Replaced several for loops with std algorithms

+++ src/correlation/test/common.hpp
@@ -39,8 +39,8 @@ std::vector<uint32_t> generate_random_vect(std::mt19937& rnd_eng, uint32_t size,
 
 std::pair<std::vector<uint32_t>, std::vector<uint32_t>> generate_correlated_vects(
     std::mt19937& rnd_eng, uint32_t size) {
-  // NOLINTNEXTLINE(readability-magic-numbers, cppcoreguidelines-avoid-magic-numbers)
-  std::uniform_int_distribution<int32_t> dist(static_cast<int32_t>(size) / -50, size / 50);
+  std::uniform_int_distribution<int32_t> dist(static_cast<int32_t>(size) / -50,  // NOLINT
+                                              static_cast<int32_t>(size / 50));  // NOLINT
   std::vector<uint32_t> v1(size);
   std::vector<uint32_t> v2(size);
   std::iota(v1.begin(), v1.end(), 0);
